+++ src/lenet_commonfunc.h
@@ -1,7 +1,7 @@
 /*
  * lenet_commonfunc.h
  *
- *  Created on: 2018Äê1ÔÂ5ÈÕ
+ *  Created on: 2018ï¿½ï¿½1ï¿½ï¿½5ï¿½ï¿½
  *      Author: admin
  */
 
@@ -10,7 +10,7 @@
 #include "lenet_define.h"
 
 
-int read_Mnist(int begin_number, int end_number, float Feature[][FEATURE_SIZE][FEATURE_SIZE], int label[]);
+int read_Mnist(int begin_number, int end_number, int Feature[][FEATURE_SIZE][FEATURE_SIZE], int label[]);
 int read_data();
 
 #endif /* LENET_COMMONFUNC_H_ */
