+++ src/lenet_matrix.h
@@ -21,11 +21,11 @@
  *      2. 由dst+=dst+src;改为dst=dst+src;(2018-1-8)
  *      3. 改为C风格(2018-1-9)
  */
-void matrix_add_24(float dst[][24], float src[][24]);
+void matrix_add_24(int dst[][24], int src[][24]);
 
-void matrix_add_10(float dst[][10], float src[][10]);
+void matrix_add_10(int dst[][10], int src[][10]);
 
-void matrix_add_8(float dst[][8], float src[][8]);
+void matrix_add_8(int dst[][8], int src[][8]);
 
 
 /*
@@ -37,18 +37,18 @@ void matrix_add_8(float dst[][8], float src[][8]);
  *      1. 创建新函数(2018-1-6)
  *      2. 改为C风格(2018-1-9)
  */
-void matrix_convolution_28_5(float picture[][28],
-        float kernel[][5],
-        float result[][24]);
+void matrix_convolution_28_5(int picture[][28],
+        int kernel[][5],
+        int result[][24]);
 
 
-void matrix_convolution_12_3(float picture[][12],
-        float kernel[][3],
-        float result[][10]);
+void matrix_convolution_12_3(int picture[][12],
+        int kernel[][3],
+        int result[][10]);
 
-void matrix_convolution_10_3(float picture[][10],
-        float kernel[][3],
-        float result[][8]);
+void matrix_convolution_10_3(int picture[][10],
+        int kernel[][3],
+        int result[][8]);
 
 /*
  * 函数名称：matrix_multi_convolution
@@ -58,13 +58,13 @@ void matrix_convolution_10_3(float picture[][10],
  * 修改记录：
  *      1. 创建新函数(2018-1-6)
  */
-void matrix_multi_convolution_12_3(float pictures[][12][12],
-        float kernel[][3][3],
-        float result[10][10]);
+void matrix_multi_convolution_12_3(int pictures[][12][12],
+        int kernel[][3][3],
+        int result[10][10]);
 
-void matrix_multi_convolution_10_3(float pictures[][10][10],
-        float kernel[][3][3],
-        float result[8][8]);
+void matrix_multi_convolution_10_3(int pictures[][10][10],
+        int kernel[][3][3],
+        int result[8][8]);
 
 /*
  * 函数名称：matrix_matrice_tanh
@@ -75,14 +75,14 @@ void matrix_multi_convolution_10_3(float pictures[][10][10],
  *      1. 创建新函数(2018-1-6)
  *      2. 改为C风格(2018-1-9)
  */
-void matrix_matrice_tanh_24(float input[][24],
-        float output[][24], float bias);
+void matrix_matrice_tanh_24(int input[][24],
+        int output[][24], int bias);
 
-void matrix_matrice_tanh_10(float input[][10],
-        float output[][10], float bias);
+void matrix_matrice_tanh_10(int input[][10],
+        int output[][10], int bias);
 
-void matrix_matrice_tanh_8(float input[][8],
-        float output[][8], float bias);
+void matrix_matrice_tanh_8(int input[][8],
+        int output[][8], int bias);
 
 /*
  * 函数名称：matrix_vector_tanh
@@ -94,9 +94,9 @@ void matrix_matrice_tanh_8(float input[][8],
  *      2. 把output[size] = ...改为output[i] = ...(2018-1-8)
  *      3. 改为C风格(2018-1-9)
  */
-void matrix_vector_tanh_128(float input[128],
-        float output[128],
-        float bias[128]);
+void matrix_vector_tanh_128(int input[128],
+        int output[128],
+        int bias[128]);
 
 /*
  * 函数名称：matrix_softmax
@@ -108,9 +108,9 @@ void matrix_vector_tanh_128(float input[128],
  *      2. 增加了softmax计算结果的输出(2018-1-8)
  *      3. 改为C风格(2018-1-9)
  */
-int matrix_softmax_10(float input[10],
-        float output[10],
-        float bias[10]);
+int matrix_softmax_10(int input[10],
+        int output[10],
+        int bias[10]);
 
 
 
@@ -123,9 +123,9 @@ int matrix_softmax_10(float input[10],
  *      1. 创建新函数(2018-1-6)
  *      3. 改为C风格(2018-1-9)
  */
-void matrix_pooling_24(float input[24][24], float output[12][12]);
+void matrix_pooling_24(int input[24][24], int output[12][12]);
 
-void matrix_pooling_8(float input[8][8], float output[4][4]);
+void matrix_pooling_8(int input[8][8], int output[4][4]);
 
 /*
  * 函数名称：matrix_MMV
@@ -136,8 +136,8 @@ void matrix_pooling_8(float input[8][8], float output[4][4]);
  *      1. 创建新函数(2018-1-6)
  *      2. temp的类型由int改为float
  */
-void matrix_MMV_256_128(float x[256], float weight[128][256], float output[128]);
+void matrix_MMV_256_128(int x[256], int weight[128][256], int output[128]);
 
-void matrix_MMV_128_10(float x[128], float weight[10][128], float output[10]);
+void matrix_MMV_128_10(int x[128], int weight[10][128], int output[10]);
 
 #endif /* LENET_MATRIX_H_ */
